enum inline fixes
