+++ packages/core/js/data_generation/rulesParser.js
@@ -1,38 +1,7 @@
 import { TestDataRules } from './testDataRules.js';
 import { SchemaParsingErrors } from './schema-parsing-errors.js';
 import { parseConstraintText } from './schema-constraint-parser.js';
-
-function startsConstraint(trimmedLine) {
-  return /^IF\s+(?:\[|\(|NOT\b)/i.test(trimmedLine);
-}
-
-function looksLikeInlineRuleSpec(ruleText) {
-  const trimmed = String(ruleText ?? '').trim();
-  if (trimmed.length === 0 || startsConstraint(trimmed)) {
-    return false;
-  }
-
-  if (
-    /^(?:enum|literal|regex|datatype\.(?:enum|literal|regex)|awd\.datatype\.(?:enum|literal|regex))\s*\(/i.test(trimmed)
-  ) {
-    return true;
-  }
-
-  if (/^(?:faker\.)?[A-Za-z][A-Za-z0-9_]*(?:\.[A-Za-z][A-Za-z0-9_]*)+(?:\s*\(.*\)\s*|\s*)$/i.test(trimmed)) {
-    return true;
-  }
-
-  if (!trimmed.includes(',')) {
-    return false;
-  }
-
-  const values = trimmed.split(',').map((value) => value.trim());
-  if (values.length < 2 || values.some((value) => value.length === 0 || value.length > 50)) {
-    return false;
-  }
-
-  return !values.some((value) => /[[\]{}()^$*+?|\\]/.test(value) || (value.includes('.') && /[A-Z]/.test(value)));
-}
+import { looksLikeInlineRuleSpec, startsConstraint } from './inline-schema-rule.js';
 
 function parseInlineRuleDefinition(line) {
   const source = String(line ?? '');
@@ -42,15 +11,18 @@ function parseInlineRuleDefinition(line) {
     }
 
     const name = source.slice(0, index).trim();
-    const rule = source.slice(index + 1).trim();
+    const rawRulePortion = source.slice(index + 1);
+    const rule = rawRulePortion.trim();
     if (name.length === 0 || !looksLikeInlineRuleSpec(rule)) {
       continue;
     }
 
+    const leadingWhitespaceLength = rawRulePortion.length - rawRulePortion.trimStart().length;
+
     return {
       name,
       rule,
-      separator: ': ',
+      separator: `:${rawRulePortion.slice(0, leadingWhitespaceLength)}`,
     };
   }
